fix license shown on gitea
The LICENSE Icon in the README was "wrong implemented".

It links to the HTML Page of the GitHub Repo. GitHub works with this feature, to get the Icon from their HTML and show it. But Gitea can only show raw images.

Now i've changed the URL to the Raw SVG of the GitHub repo.

Also i have changed the Link which will opened on click. It opens the link on github if user watches the readme on github, and the same for gitea and also local markdown viewers.